+++ lib/bower-rails.rb
@@ -7,6 +7,9 @@ module BowerRails
   class << self
     # An array of tasks to enhance `rake assets:precompile`
     attr_reader :tasks
+
+    # If set to true then rake bower:install task is invoked before assets precompilation
+    attr_accessor :install_before_precompile
     
     # If set to true then rake bower:install && rake bower:resolve tasks
     # are invoked before assets precompilation
@@ -24,6 +27,7 @@ module BowerRails
     private
 
       def collect_tasks
+        @tasks << ['bower:install'] if @install_before_precompile
         @tasks << ['bower:install', 'bower:resolve'] if @resolve_before_precompile
         @tasks << ['bower:install', 'bower:clean']   if @clean_before_precompile
         @tasks.flatten!
@@ -35,6 +39,7 @@ module BowerRails
   @tasks = []
 
   # Set default values for options
+  @install_before_precompile = false
   @resolve_before_precompile = false
-  @clean_before_precompile = false
+  @clean_before_precompile   = false
 end