+++ lib/generators/bower_rails/initialize/templates/bower_rails.rb
@@ -1,4 +1,7 @@
 BowerRails.configure do |bower_rails|
+  # Invoke `rake bower:install` automatically before the `rake assets:precompile` task.
+  # bower_rails.install_before_precompile = true
+
   # Invoke `rake bower:resolve` automatically before the `rake assets:precompile` task.
   # bower_rails.resolve_before_precompile = true